% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Updated_Graph.T2.test.R
\name{graph.T2.test}
\alias{graph.T2.test}
\title{Performs the Hotelling T2 test in Fourier space.
Modified based on the 'graph.T2.test' function from R package 'DEGraph'.
Updates include: (1) updated the tolerance for detecting linear dependencies;
(2) added an argument "nmin", the minimum of sample size between the two groups.}
\usage{
graph.T2.test(X1, X2, G = NULL, lfA = NULL, k = ncol(X1), nmin)
}
\arguments{
\item{X1}{A n1 x p numeric matrix, observed data for class 1: p variables, n1 observations.}

\item{X2}{A n2 x p numeric matrix, observed data for class 2: p variables, n2 observations.}

\item{G}{An object of class graphAM or graphNEL, the graph to be used in the two-sample test.}

\item{lfA}{A list returned by laplacianFromA(), containing the Laplacian eigen vectors and eigen values.}

\item{k}{A numeric value, number of Fourier components retained for the test.}

\item{nmin}{the minimum of sample size between the two groups.}

\item{alpha0}{significance level of the hypothesis test.}
}
\value{
A list with class "htest".
}
\description{
Performs the Hotelling T2 test in Fourier space.
Modified based on the 'graph.T2.test' function from R package 'DEGraph'.
Updates include: (1) updated the tolerance for detecting linear dependencies;
(2) added an argument "nmin", the minimum of sample size between the two groups.
}
