% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2DAG.R
\name{SEM.fit}
\alias{SEM.fit}
\title{Fit the SEM models used for estimating the coefficient matrix Q.}
\usage{
SEM.fit(p.model, ntotal, prefix, Data, A)
}
\arguments{
\item{p.model}{the # of regressions to run, which is equal to the number of variables (genes)
that are children nodes of at least one other variable in the directed acyclic graph
that describes the pathway.}

\item{ntotal}{The total sample size n1 + n2}

\item{prefix}{The prefix of the column names in the data matrix 'Data'}

\item{Data}{A ntotal x p data frame, the first n1 rows store the standardized data
of the n1 samples from group 1, the second n1 rows store the standardized data
of the n2 samples from group 2.}

\item{A}{The binary adjacency matrix, where an entry aij=0 indicates no directed relationship
from variable i to variable j, and aij=1 indicates that there exists a directed relationship
from variable i to variable j.}

\item{Qest}{The estimate of the coefficient matrix Q.}

\item{Rdiag}{The estimate of the covariance matrix of the residuals.}
}
\value{
A list that contains the following elements:
}
\description{
Fit the SEM models used for estimating the coefficient matrix Q.
}
