% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2DAG.R
\name{T2DAG}
\alias{T2DAG}
\title{Performs the T2DAG test.}
\usage{
T2DAG(X, Y, A)
}
\arguments{
\item{X}{The n1 x p sample matrix from group 1. Each row corresponds to a p-dimensional sample.}

\item{Y}{The n2 x p sample matrix from group 2. Each row corresponds to a p-dimensional sample.}

\item{A}{The binary adjacency matrix, where an entry aij=0 indicates no directed relationship
from variable i to variable j, and aij=1 indicates that there exists a directed relationship
from variable i to variable j.}

\item{Qest}{The estimate of the coefficient matrix Q.}

\item{Rdiag}{The estimate of the covariance matrix of the residuals.}

\item{T2DAG.stat}{T2DAG test statistic.}

\item{T2DAG.pval}{Conclusion of the T2DAG test. 1 if H0 rejected and 0 otherwise.}
}
\value{
A list that contains the following elements:
}
\description{
Performs the T2DAG test.
}
